+++ src/components/survey/index.ts
@@ -1,0 +1,6 @@
+export { default as Step1 } from './steps/Step1';
+export { default as Step2 } from './steps/Step2';
+export { default as Step3 } from './steps/Step3';
+export { default as Step4 } from './steps/Step4';
+export { default as Step5 } from './steps/Step5';
+export { default as LoadingScreen } from './LoadingScreen/LoadingScreen';
